#19 modularises routes and cleans server.js
